+++ test/conformance/navigation_event_test.exs
@@ -6,6 +6,7 @@ defmodule Fluffy.Conformance.NavigationEventTest do
   import Fluffy.Locator
 
   alias Fluffy.Event
+  alias Fluffy.Navigation
   alias Fluffy.NavigationEvent
   alias Fluffy.TestHTTPFixtures
 
@@ -28,9 +29,9 @@ defmodule Fluffy.Conformance.NavigationEventTest do
         session
         |> visit(TestHTTPFixtures.path(fixture, "/start"))
         |> wait_for(Event.navigation(:continue), &click(&1, by_role(:link, name: "Continue")))
-        |> expect(to_have_navigation_from_url(:continue, from_url))
-        |> expect(to_have_navigation_url(:continue, destination))
-        |> expect(to_have_navigation_status(:continue, 202))
+        |> expect(Navigation.to_have_from_url(:continue, from_url))
+        |> expect(Navigation.to_have_url(:continue, destination))
+        |> expect(Navigation.to_have_status(:continue, 202))
         |> expect("Arrived" |> by_text() |> to_be_visible())
 
       assert %NavigationEvent{url: ^destination, status: 202} =
@@ -65,8 +66,8 @@ defmodule Fluffy.Conformance.NavigationEventTest do
       |> visit(TestHTTPFixtures.path(fixture, "/start"))
       |> fill(by_label("Query"), "fluffy")
       |> wait_for(Event.navigation(:submit), &click(&1, by_role(:button, name: "Search")))
-      |> expect(to_have_navigation_url(:submit, destination))
-      |> expect(to_have_navigation_status(:submit, 201))
+      |> expect(Navigation.to_have_url(:submit, destination))
+      |> expect(Navigation.to_have_status(:submit, 201))
       |> expect("Submitted" |> by_text() |> to_be_visible())
 
       [request] =
@@ -95,8 +96,8 @@ defmodule Fluffy.Conformance.NavigationEventTest do
       session
       |> visit(TestHTTPFixtures.path(fixture, "/start"))
       |> wait_for(Event.navigation(:redirect), &click(&1, by_role(:link, name: "Continue")))
-      |> expect(to_have_navigation_url(:redirect, destination))
-      |> expect(to_have_navigation_status(:redirect, 203))
+      |> expect(Navigation.to_have_url(:redirect, destination))
+      |> expect(Navigation.to_have_status(:redirect, 203))
       |> expect("Final" |> by_text() |> to_be_visible())
     end
   end